配置

+++ src/main/java/com/mytest/admin/po/THistoryInfoPo.java
@@ -41,7 +41,7 @@ public class THistoryInfoPo extends MBeanBase implements Serializable {
 	@Column(name="regiest_num")
 	private int regiestNum;
 	@Column(name="regiest_num_rate")
-	private int regiestNumRate;
+	private double regiestNumRate;
 	@Column(name="use_tickt_num")
 	private int useTicktNum;
 	@Column(name="createtime")
